Add a proper interface to flashes service
Before this commit the only way to add a flash message to the flashes
service was to use `loadFlashes` function, which is not a greate API for
a service. In order to make it more intuitive I added 3 functions:
`error`, `notice` and `warning`, each taking only one argument: a
message that you want to show.