WIP with passing tests but incorrect behavior
This would include incorrect data, as we are not filtering by org membership.  So if you were to visit
rails/rails, it would begin to show up in your sidebar, which is undesired.

Blocking until we have a better v3 way of expressing this, i.e. with push permissions attached to the repo
records